lambda_pipeline/requirements.txt
pyspark>=3.3.0
pandas>=1.3.0
numpy>=1.21.0
kafka-python>=2.0.2
confluent-kafka>=1.8.2
pyarrow>=6.0.0
jupyter>=1.0.0
python-dotenv>=0.19.0
